+++ conductor-client/src/main/kotlin/com/openlattice/organizations/ExternalDatabaseManagementService.kt
@@ -14,7 +14,6 @@ import com.openlattice.authorization.securable.SecurableObjectType
 import com.openlattice.edm.PropertyTypeIdFqn
 import com.openlattice.edm.processors.GetEntityTypeFromEntitySetEntryProcessor
 import com.openlattice.edm.processors.GetFqnFromPropertyTypeEntryProcessor
-import com.openlattice.edm.processors.GetOrganizationIdFromEntitySetEntryProcessor
 import com.openlattice.edm.requests.MetadataUpdate
 import com.openlattice.edm.set.EntitySetFlag
 import com.openlattice.hazelcast.HazelcastMap
@@ -32,7 +31,6 @@ import com.openlattice.postgres.PostgresColumn.*
 import com.openlattice.postgres.ResultSetAdapters.*
 import com.openlattice.postgres.external.ExternalDatabaseConnectionManager
 import com.openlattice.postgres.external.ExternalDatabasePermissioningService
-import com.openlattice.postgres.mapstores.SecurableObjectTypeMapstore
 import com.openlattice.postgres.streams.BasePostgresIterable
 import com.openlattice.postgres.streams.PreparedStatementHolderSupplier
 import com.openlattice.postgres.streams.StatementHolderSupplier
@@ -76,9 +74,6 @@ class ExternalDatabaseManagementService(
     private val primaryKeyConstraint = "PRIMARY KEY"
     private val FETCH_SIZE = 100_000
 
-    /**
-     * Only needed for materialize entity set, which should move elsewhere eventually
-     */
     private val entitySets = HazelcastMap.ENTITY_SETS.getMap(hazelcastInstance)
     private val propertyTypes = HazelcastMap.PROPERTY_TYPES.getMap(hazelcastInstance)
     private val transporterState = HazelcastMap.TRANSPORTER_DB_COLUMNS.getMap(hazelcastInstance)
@@ -295,6 +290,10 @@ class ExternalDatabaseManagementService(
         }.toList()
     }
 
+    fun executePrivilegesUpdate(action: Action, acls: List<Acl>) {
+        extDbPermsManager.executePrivilegesUpdate(action, acls)
+    }
+
     /*UPDATE*/
     fun updateOrganizationExternalDatabaseTable(orgId: UUID, tableFqnToId: Pair<String, UUID>, update: MetadataUpdate) {
         update.name.ifPresent {
@@ -460,77 +459,6 @@ class ExternalDatabaseManagementService(
     fun getOrganizationOwners(orgId: UUID): Set<SecurablePrincipal> {
         val principals = securePrincipalsManager.getAuthorizedPrincipalsOnSecurableObject(AclKey(orgId), EnumSet.of(Permission.OWNER))
         return securePrincipalsManager.getSecurablePrincipals(principals).toSet()
-    }
-
-    /**
-     * Sets privileges for a user on an organization's column
-     */
-
-    fun executePrivilegesUpdate(action: Action, acls: List<Acl>) {
-        val aclsByType = mapAclsToExternalVsAssemblyColumnAcls(acls)
-
-        // for entityset aclkeys:
-        val assemblyCols = aclsByType.getValue(SecurableObjectType.PropertyTypeInEntitySet)
-        executePrivilegesUpdateOnPropertyTypes(action, assemblyCols)
-
-        // for organizationexternalDatabaseColumns:
-        val externalTableColAcls = aclsByType.getValue(SecurableObjectType.OrganizationExternalDatabaseColumn)
-        executePrivilegesUpdateOnOrgExternalDbColumns(action, externalTableColAcls)
-    }
-
-    fun executePrivilegesUpdateOnOrgExternalDbColumns(action: Action, externalTableColAcls: List<Acl>) {
-        val extTableColIds = externalTableColAcls.aclKeysAsSet {
-            it.aclKey[1]
-        }
-        val columnsById = organizationExternalDatabaseColumns.getAll(extTableColIds).values.associate {
-            AclKey( it.tableId, it.id) to TableColumn(it.organizationId, it.tableId, it.id)
-        }
-        extDbPermsManager.updateExternalTablePermissions(action, externalTableColAcls, columnsById)
-    }
-
-    fun executePrivilegesUpdateOnPropertyTypes(action: Action, assemblyAcls: List<Acl> ){
-        val esids = assemblyAcls.aclKeysAsSet {
-            it.aclKey[0]
-        }
-        val aclKeyToTableCols = entitySets.submitToKeys(esids, GetOrganizationIdFromEntitySetEntryProcessor()).thenApplyAsync { esidToOrgId ->
-            assemblyAcls.associate {
-                it.aclKey to TableColumn(esidToOrgId.getValue(it.aclKey[0]), it.aclKey[0], it.aclKey[1])
-            }
-        }
-        extDbPermsManager.updateAssemblyPermissions(action, assemblyAcls, aclKeyToTableCols.toCompletableFuture().get())
-    }
-
-    private fun mapAclsToExternalVsAssemblyColumnAcls(acls: List<Acl>): Map<SecurableObjectType, List<Acl>>{
-        val aclKeyIndex = acls.aclKeysAsSet { it.aclKey.index }.toTypedArray()
-
-        val allOrgExternalDBAclKeys: Set<AclKey> = securableObjectTypes.keySet(
-                getAclKeysOfObjectTypePredicate( aclKeyIndex, SecurableObjectType.OrganizationExternalDatabaseColumn)
-        )
-        val allOrgExternalDbAcls = acls.filter { allOrgExternalDBAclKeys.contains(it.aclKey) }
-
-        val allAssemblyAclKeys: Set<AclKey> = securableObjectTypes.keySet(
-                getAclKeysOfObjectTypePredicate( aclKeyIndex, SecurableObjectType.PropertyTypeInEntitySet)
-        )
-        val allAssemblyAcls = acls.filter { allAssemblyAclKeys.contains(it.aclKey) }
-
-        return mapOf(
-                SecurableObjectType.OrganizationExternalDatabaseColumn to allOrgExternalDbAcls,
-                SecurableObjectType.PropertyTypeInEntitySet to allAssemblyAcls
-        )
-    }
-
-    private fun getAclKeysOfObjectTypePredicate(
-            aclKeysIndexForm: Array<String>,
-            objectType: SecurableObjectType
-    ): Predicate<AclKey, SecurableObjectType> {
-        return Predicates.and<AclKey, SecurableObjectType>(
-                Predicates.`in`<Any, Any>(SecurableObjectTypeMapstore.ACL_KEY_INDEX, *aclKeysIndexForm),
-                Predicates.equal<Any, Any>(SecurableObjectTypeMapstore.SECURABLE_OBJECT_TYPE_INDEX, objectType )
-        )
-    }
-
-    private fun <T> Collection<Acl>.aclKeysAsSet( extraction: ( Acl ) -> T ): Set<T> {
-        return this.mapTo( HashSet(this.size) ) { extraction(it) }
     }
 
     /**
